+++ app/src/main/java/com/example/mybudget1/SpentItem.java
@@ -2,10 +2,10 @@ package com.example.mybudget1;
 
 public class SpentItem {
     private String name;
-    private int amount;
+    private double amount;
     private String date;
 
-    public SpentItem(String name, int amount, String date) {
+    public SpentItem(String name, double amount, String date) {
         this.name = name;
         this.amount = amount;
         this.date = date;
@@ -15,14 +15,14 @@ public class SpentItem {
         return name;
     }
 
-    public int getAmount() {return amount;}
+    public double getAmount() {return amount;}
 
     public String getDate() {
         return date;
     }
 
 
-    public void change(String newName , int newAmount , String newDate){
+    public void change(String newName , double newAmount , String newDate){
         this.name = newName;
         this.amount = newAmount;
         this.date = newDate;
